speech-to-speech/workshops/python-server/requirements.txt
boto3==1.38.2
numpy==2.2.4
scipy==1.15.2
rx==3.2.0
websockets==15.0.1
aws_sdk_bedrock_runtime
mcp==1.8.1
strands-agents